+++ janus/janus-demos/simulation-demos/boids/src/main/java/org/janusproject/demos/simulation/boids/osgi/BoidsActivator.java
@@ -24,17 +24,16 @@ import java.awt.Color;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import javax.vecmath.Vector2d;
-
-import org.arakhne.vmutil.locale.Locale;
+import org.arakhne.afc.math.continous.object2d.Vector2f;
+import org.arakhne.afc.vmutil.locale.Locale;
 import org.janusproject.demos.simulation.boids.Launcher;
 import org.janusproject.demos.simulation.boids.agent.BoidAgent;
 import org.janusproject.demos.simulation.boids.agent.EnvironmentAgent;
 import org.janusproject.demos.simulation.boids.util.EnvironmentFrame;
 import org.janusproject.demos.simulation.boids.util.Population;
 import org.janusproject.demos.simulation.boids.util.Settings;
-import org.janusproject.kernel.address.AgentAddress;
 import org.janusproject.kernel.Kernel;
+import org.janusproject.kernel.address.AgentAddress;
 import org.janusproject.kernel.agent.KernelAgentFactory;
 import org.janusproject.kernel.logger.LoggerUtil;
 import org.janusproject.kernel.mmf.JanusApplication;
@@ -133,8 +132,8 @@ public class BoidsActivator  implements BundleActivator, JanusApplication {
 
 	
 	private static AgentAddress addBoid(Kernel kernel, Population p, String name) {
-		Vector2d initialPosition = new Vector2d((Math.random() - 0.5)*Settings.ENVIRONMENT_DEMI_WIDTH, (Math.random() - 0.5)*Settings.ENVIRONMENT_DEMI_HEIGHT);
-		Vector2d initialSpeed = new Vector2d(Math.random() - 0.5, Math.random() - 0.5);
+		Vector2f initialPosition = new Vector2f((Math.random() - 0.5)*Settings.ENVIRONMENT_DEMI_WIDTH, (Math.random() - 0.5)*Settings.ENVIRONMENT_DEMI_HEIGHT);
+		Vector2f initialSpeed = new Vector2f(Math.random() - 0.5, Math.random() - 0.5);
 		
 		BoidAgent boid = new BoidAgent(p, initialPosition, initialSpeed);
